rename GlobalSnackbar to GlobalSnackbarProvider
