+++ src/ui/menus/CommandPalette/ListItems/Import.tsx
@@ -1,5 +1,5 @@
 import { UploadFile } from '@mui/icons-material';
-import { useGlobalSnackbar } from '../../../../components/GlobalSnackbar';
+import { useGlobalSnackbar } from '../../../../components/GlobalSnackbarProvider';
 import { CSV_IMPORT_MESSAGE } from '../../../../constants/appConstants';
 import { CommandPaletteListItem, CommandPaletteListItemSharedProps } from '../CommandPaletteListItem';
